- Moved FgProfile into it's own directory within /pages - Created FgProfileService to handle backend calls for FgProfile - Calling the service from App.js is causing an error when trying to access member.fullName() function within the FgProfile component. This doesn't happen when passing in the mocked member directly. (Investigate) - Kept FgProfileService file and calls within App.js but reverted back to passing in the Mocked type directly.
