+++ App.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import { View } from 'react-native';
-import { FgProfile } from "./src/pages/FgProfile";
+import { FgProfile } from "./src/pages/FgProfile/FgProfile";
+import { FgProfileService } from "./src/pages/FgProfile/FgProfile-Service";
 import { AppLoading, Font } from 'expo'
 import {
-    MOCKED_MEMBER_DARIA_with_BANNER_and_AVATAR, MOCKED_MEMBER_DARIA_with_BANNER_no_AVATAR
+    MOCKED_MEMBER_DARIA_with_AVATAR_no_BANNER, MOCKED_MEMBER_DARIA_with_BANNER_and_AVATAR,
+    MOCKED_MEMBER_DARIA_with_BANNER_no_AVATAR
 } from "./src/test/MockedTypes";
+import {FgMember} from "./src/types/FgMember";
+
+const MOCKED_MEMBER_ID = '77bac705-1ce2-4a7b-8247-cf4bdfd451a9';
 
 export default class App extends React.Component {
+
+    service = new FgProfileService();
+    member : FgMember = this.service.getMemberWithId(MOCKED_MEMBER_ID);
 
     state = {
         isReady: false
@@ -30,7 +38,7 @@ export default class App extends React.Component {
 
         return (
             <View>
-                <FgProfile member={ MOCKED_MEMBER_DARIA_with_BANNER_no_AVATAR}/>
+                <FgProfile member={ MOCKED_MEMBER_DARIA_with_BANNER_no_AVATAR }/>
             </View>
         );
     }
